refactor interdace in 5.superservice
