+++ course3/3.clean_architecture/5.super_service/task3.3.5.1/internal/controller/responder.go
@@ -6,14 +6,6 @@ import (
 	"net/http"
 )
 
-type Responder interface {
-	OutputJSON(w http.ResponseWriter, data any)
-
-	ErrorBadRequest(w http.ResponseWriter, err error)
-	ErrorInternalServerError(w http.ResponseWriter, err error)
-	ErrorNotFound(w http.ResponseWriter)
-}
-
 type ResponderImpl struct {
 	logger *log.Logger
 }
